reafactor: move health query to margin account
