+++ dango/account-margin/Cargo.toml
@@ -21,6 +21,7 @@ library = []
 anyhow        = { workspace = true }
 dango-auth    = { workspace = true }
 dango-lending = { workspace = true, features = ["library"] }
+dango-oracle  = { workspace = true, features = ["library"] }
 dango-types   = { workspace = true }
 grug          = { workspace = true }
