+++ dango/account-margin/src/execute.rs
@@ -1,15 +1,10 @@
 use {
+    crate::query_health,
     anyhow::ensure,
     dango_auth::authenticate_tx,
-    dango_lending::{calculate_account_health, DEBTS},
-    dango_types::{
-        account::InstantiateMsg,
-        config::{ACCOUNT_FACTORY_KEY, LENDING_KEY},
-        lending::LendingAppConfig,
-    },
+    dango_types::{account::InstantiateMsg, config::ACCOUNT_FACTORY_KEY},
     grug::{
-        Addr, AuthCtx, AuthResponse, BorshDeExt, Coins, MutableCtx, NumberConst, Response,
-        StdResult, Tx, Udec128,
+        Addr, AuthCtx, AuthResponse, MutableCtx, NumberConst, Response, StdResult, Tx, Udec128,
     },
 };
 
@@ -35,23 +30,7 @@ pub fn authenticate(ctx: AuthCtx, tx: Tx) -> anyhow::Result<AuthResponse> {
 
 #[cfg_attr(not(feature = "library"), grug::export)]
 pub fn backrun(ctx: AuthCtx, _tx: Tx) -> anyhow::Result<Response> {
-    let lending_config: LendingAppConfig = ctx.querier.query_app_config(LENDING_KEY)?;
-
-    // Query all debts for the account.
-    let debts = ctx
-        .querier
-        .query_wasm_raw(lending_config.lending, DEBTS.path(ctx.contract))?
-        .map(|coins| coins.deserialize_borsh::<Coins>())
-        .transpose()?
-        .unwrap_or_default();
-
-    // Calculate the utilization rate.
-    let health = calculate_account_health(
-        &ctx.querier,
-        ctx.contract,
-        debts,
-        lending_config.collateral_powers,
-    )?;
+    let health = query_health(ctx.into_immutable())?;
 
     // If the utilization rate is greater than 1, the account is undercollateralized.
     ensure!(
