+++ dango/account-spot/src/query.rs
@@ -1,6 +1,6 @@
 use {
     dango_auth::NEXT_SEQUENCE,
-    dango_types::account::single::QueryMsg,
+    dango_types::account::spot::QueryMsg,
     grug::{ImmutableCtx, Json, JsonSerExt, StdResult, Storage},
 };
