+++ dango/types/src/lending.rs
@@ -84,17 +84,6 @@ pub enum ExecuteMsg {
     Repay {},
 }
 
-#[grug::derive(Serde)]
-pub struct HealthResponse {
-    /// The margin account's utilization rate.
-    pub utilization_rate: Udec128,
-    /// The total value of the margin account's debt.
-    pub total_debt_value: Udec128,
-    /// The total value of the margin account's collateral, adjusted for
-    /// the collateral power of each denom.
-    pub total_adjusted_collateral_value: Udec128,
-}
-
 #[grug::derive(Serde, QueryRequest)]
 pub enum QueryMsg {
     /// Query the lending market of a single token.
@@ -118,7 +107,4 @@ pub enum QueryMsg {
     /// Returns all collateral powers.
     #[returns(BTreeMap<Denom, CollateralPower>)]
     CollateralPowers {},
-    /// Queries the health of a margin account.
-    #[returns(HealthResponse)]
-    Health { account: Addr },
 }
